requirements.txt
kafka-python
pandas
SQLAlchemy
numpy
scikit-learn
xgboost
joblib
Flask-Login
pymysql